--- src/main/java/UnknownCommand.java
@@ -1,9 +1,0 @@
-public class UnknownCommand extends Command{
-
-    public UnknownCommand() {
-    }
-
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
-        ui.showError("OOPS!!! I'm sorry, but I don't know what that means :-(");
-    }
-}
