+++ src/main/java/duke/command/AddCommand.java
@@ -1,2 +1,4 @@
+package duke.command;
+
 public abstract class AddCommand extends Command {
 }
