+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
+import duke.task.Deadline;
 import java.time.LocalDate;
 
 public class DeadlineCommand extends Command {
@@ -11,7 +17,7 @@ public class DeadlineCommand extends Command {
         this.byWhen = byWhen;
     }
 
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
             Deadline newTask = new Deadline(desc,byWhen);
             tasks.addTask(newTask);
     }
