+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,4 +1,10 @@
-public class DeleteCommand extends AddCommand{
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
+
+public class DeleteCommand extends AddCommand {
     public static final String COMMAND_WORD = "delete";
     private int taskIndex;
 
@@ -6,7 +12,7 @@ public class DeleteCommand extends AddCommand{
         this.taskIndex = taskIndex;
     }
 
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.deleteTask(taskIndex);
     }
 }
