+++ src/main/java/duke/command/MarkCommand.java
@@ -1,4 +1,10 @@
-public class MarkCommand extends Command{
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
+
+public class MarkCommand extends Command {
     public static final String COMMAND_WORD = "mark";
     private int taskIndex;
 
@@ -6,7 +12,7 @@ public class MarkCommand extends Command{
         this.taskIndex = taskIndex;
     }
 
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.markTask(taskIndex);
     }
 }
