+++ src/main/java/duke/command/TerminateCommand.java
@@ -1,9 +1,14 @@
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
 public class TerminateCommand extends Command {
     public static final String COMMAND_WORD = "bye";
 
     public TerminateCommand() {
     }
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         storage.saveTasks(tasks.getArrayList());
         ui.showGoodBye();
     }
