+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
+import duke.task.Todo;
+
 public class ToDoCommand extends AddCommand {
     public static final String COMMAND_WORD = "todo";
     private String desc;
@@ -6,7 +13,7 @@ public class ToDoCommand extends AddCommand {
         this.desc = desc;
     }
 
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         Todo newTask = new Todo(desc);
         tasks.addTask(newTask);
     }
