+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,4 +1,9 @@
-public class UnmarkCommand extends Command{
+package duke.command;
+
+import duke.Ui;
+import duke.TaskList;
+import duke.Storage;
+public class UnmarkCommand extends Command {
     public static final String COMMAND_WORD = "unmark";
     private int taskIndex;
 
@@ -6,7 +11,7 @@ public class UnmarkCommand extends Command{
         this.taskIndex = taskIndex;
     }
 
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.unmarkTask(taskIndex);
     }
 }
