+++ src/main/java/duke/task/Event.java
@@ -1,7 +1,9 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Event extends Task {
     LocalDate from;
     LocalDate to;
