+++ text-ui-test/EXPECTED.TXT
@@ -5,23 +5,23 @@ Hello from
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 How may I be of service to you?
-Enter your command:
-Got it. I've added this task:
+Enter your duke.command:
+Got it. I've added this duke.task:
 [T][ ] borrow book
-Now you have 1 task(s) in the list.
-Got it. I've added this task:
+Now you have 1 duke.task(s) in the list.
+Got it. I've added this duke.task:
 [D][ ] return book [by June 6th]
-Now you have 2 task(s) in the list.
-Got it. I've added this task:
+Now you have 2 duke.task(s) in the list.
+Got it. I've added this duke.task:
 [E][ ] project meeting [from: Aug 6th 2pm to: 4pm]
-Now you have 3 task(s) in the list.
-Nice! I've marked this task as done:
+Now you have 3 duke.task(s) in the list.
+Nice! I've marked this duke.task as done:
  [T][X] borrow book
-Nice! I've marked this task as done:
+Nice! I've marked this duke.task as done:
  [E][X] project meeting [from: Aug 6th 2pm to: 4pm]
-Nice! I've marked this task as done:
+Nice! I've marked this duke.task as done:
  [D][X] return book [by June 6th]
-OK, I've marked this task as not done yet:
+OK, I've marked this duke.task as not done yet:
  [T][ ] borrow book
 Here are the tasks in your list:
 1. [T][ ] borrow book
